fix(ci): disable host dns test on ci

+++ scripts/dns/start_dns.sh
@@ -62,29 +62,41 @@ rm -rf $coredns_config_dir && mkdir -p $coredns_config_dir
 # Replace the placeholder in Corefile.template with the actual etcd IP, and write it to config directory.
 template=$(cat ./assets/Corefile.template)
 echo "${template/xxxxxx/$etcd_ip}" >$coredns_config_dir/Corefile
-# Disable systemd-resolved.
-systemctl stop systemd-resolved &> /dev/null
-systemctl disable systemd-resolved &> /dev/null
-cat << EOF
+# Disable systemd-resolved (when not on ci).
+if [[ $KUBE_CI_MODE != "ON" ]]; then
+	systemctl stop systemd-resolved &>/dev/null
+	systemctl disable systemd-resolved &>/dev/null
+	cat <<EOF
 systemd-resolved has been disabled on this machine to free up udp 53 port. 
 To use our CoreDNS name server, modify /etc/resolv.conf and append 
-    nameserver 127.0.0.1
+	nameserver 127.0.0.1
 If you would like to turn systemd-resolved back on (they can be automatically
 restarted with stop_dns.sh), please shutdown CoreDNS nameserver, then type
-    systemctl enable systemd-resolved
-    systemctl start systemd-resolved
+	systemctl enable systemd-resolved
+	systemctl start systemd-resolved
 EOF
+fi
 # Start CoreDNS container.
 docker start $coredns_container_name &>/dev/null
 if [ $? -ne 0 ]; then
-	docker run -d \
-		--name $coredns_container_name \
-		--restart always \
-		-v $coredns_config_dir:/etc/coredns \
-		-p 53:53/udp \
-		coredns/coredns:1.9.1 \
-		-conf /etc/coredns/Corefile &>/dev/null &&
-		echo "CoreDNS container started, name is ${coredns_container_name}"
+	if [[ $KUBE_CI_MODE != "ON" ]]; then
+		docker run -d \
+			--name $coredns_container_name \
+			--restart always \
+			-v $coredns_config_dir:/etc/coredns \
+			-p 53:53/udp \
+			coredns/coredns:1.9.1 \
+			-conf /etc/coredns/Corefile &>/dev/null &&
+			echo "CoreDNS container started, name is ${coredns_container_name}"
+	else
+		docker run -d \
+			--name $coredns_container_name \
+			--restart always \
+			-v $coredns_config_dir:/etc/coredns \
+			coredns/coredns:1.9.1 \
+			-conf /etc/coredns/Corefile &>/dev/null &&
+			echo "CoreDNS container started, name is ${coredns_container_name}"
+	fi
 else
 	echo "CoreDNS container already started"
 fi
@@ -93,6 +105,6 @@ fi
 get_container_ip $coredns_container_name
 write_component_ip "coredns" $ip
 
-# Write CoreDNS's node IP to etcd, so that host machines can also 
+# Write CoreDNS's node IP to etcd, so that host machines can also
 # access name server by modifying /etc/resolv.conf.
 write_component_ip "coredns-host" $apiserver_ip
